Start moving towards better API design.
Started thinking about splitting env into it's own model, with the idea
that someone defines the environments available at a company. Then the
front end can query this if someone needs to know the available
environments.
Along with this we'll need more endpoints and a cleaner way to query the
models.
Also think about the way secrets will be updated on a project. Some sort
of logic needs to decide if the api call can continue.